Move mentions out of LokiAPI
