--- java/src/main/java/org/whispersystems/signalservice/loki/messaging/Mention.kt
@@ -1,3 +1,0 @@
-package org.whispersystems.signalservice.loki.messaging
-
-data class Mention(val hexEncodedPublicKey: String, val displayName: String)
