+++ java/src/main/java/org/whispersystems/signalservice/loki/messaging/mentions/Mention.kt
@@ -1,0 +1,3 @@
+package org.whispersystems.signalservice.loki.messaging.mentions
+
+data class Mention(val hexEncodedPublicKey: String, val displayName: String)
